fix: fix compile error and segfault
* fix(Client): remove double free

* refactor(AMethod): refactor path recombination

* fix(LocationConfig): change default route to '/'

* fix(Server): remove unnecessary type conversion

+++ srcs/server/src/Server.cpp
@@ -60,7 +60,7 @@ void Server::hostInit(void) {
   std::list<IServerConfig *> serverInfo = config.getServerConfigs();
   std::list<IServerConfig *>::iterator it = serverInfo.begin();
   while (it != serverInfo.end()) {
-    if ((*it)->getListen() == (size_t)this->_port) {
+    if ((*it)->getListen() == this->_port) {
       this->_host = (*it)->getServerName();
       return;
     } else {
